refactor(lungo): extract TransportRail into its own file
Move the TransportRail component and its helpers out of CustomNode.tsx
into TransportRail.tsx to stay under the 400-line lint threshold.
